# backend: optimize updating command metrics