--- a/pkg/proxy/backend/metrics.go
+++ b/pkg/proxy/backend/metrics.go
@@ -15,14 +15,14 @@ import (
 
 type cmdMetricsCache struct {
 	sync.Mutex
-	counter   map[string][pnet.ComEnd]prometheus.Counter  // addr -> [cmd]counter
-	histogram map[string][pnet.ComEnd]prometheus.Observer // addr -> [cmd]observer
+	counter   map[string]*[pnet.ComEnd]prometheus.Counter  // addr -> [cmd]counter
+	histogram map[string]*[pnet.ComEnd]prometheus.Observer // addr -> [cmd]observer
 }
 
 func newCmdMetricsCache() cmdMetricsCache {
 	return cmdMetricsCache{
-		counter:   make(map[string][pnet.ComEnd]prometheus.Counter),
-		histogram: make(map[string][pnet.ComEnd]prometheus.Observer),
+		counter:   make(map[string]*[pnet.ComEnd]prometheus.Counter),
+		histogram: make(map[string]*[pnet.ComEnd]prometheus.Observer),
 	}
 }
 
@@ -34,7 +34,7 @@ func addCmdMetrics(cmd pnet.Command, addr string, startTime monotime.Time) {
 
 	addrCounter, ok := cache.counter[addr]
 	if !ok {
-		addrCounter = [pnet.ComEnd]prometheus.Counter{}
+		addrCounter = &[pnet.ComEnd]prometheus.Counter{}
 		cache.counter[addr] = addrCounter
 	}
 	counter := addrCounter[cmd]
@@ -46,10 +46,9 @@ func addCmdMetrics(cmd pnet.Command, addr string, startTime monotime.Time) {
 
 	// The duration labels are different with TiDB: Labels in TiDB are statement types.
 	// However, the proxy is not aware of the statement types, so we use command types instead.
-	cost := monotime.Since(startTime)
 	addrHist, ok := cache.histogram[addr]
 	if !ok {
-		addrHist = [pnet.ComEnd]prometheus.Observer{}
+		addrHist = &[pnet.ComEnd]prometheus.Observer{}
 		cache.histogram[addr] = addrHist
 	}
 	hist := addrHist[cmd]
@@ -57,6 +56,7 @@ func addCmdMetrics(cmd pnet.Command, addr string, startTime monotime.Time) {
 		hist = metrics.QueryDurationHistogram.WithLabelValues(addr, cmd.String())
 		addrHist[cmd] = hist
 	}
+	cost := monotime.Since(startTime)
 	hist.Observe(cost.Seconds())
 }