--- a/pkg/proxy/backend/metrics_test.go
+++ b/pkg/proxy/backend/metrics_test.go
@@ -9,8 +9,18 @@ import (
 
 	"github.com/pingcap/tiproxy/pkg/metrics"
 	pnet "github.com/pingcap/tiproxy/pkg/proxy/net"
+	"github.com/pingcap/tiproxy/pkg/util/monotime"
 	"github.com/prometheus/client_golang/prometheus"
 )
+
+func BenchmarkAddCmdMetrics(b *testing.B) {
+	cmd := pnet.ComQuery
+	addr := "127.0.0.1:4000"
+	startTime := monotime.Now()
+	for i := 0; i < b.N; i++ {
+		addCmdMetrics(cmd, addr, startTime)
+	}
+}
 
 func BenchmarkAddCmdMetricsArrayPointer(b *testing.B) {
 	cmd := pnet.ComQuery