<!--

Thank you for contributing to TiProxy!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close: #xxx" or "ref: #xxx".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close #420

Problem Summary:
When running sysbench point select, the CPU cost of adding metrics of CPS and duration per cmd is about 3.6%.

What is changed and how it works:
- Cache the results of `WithLabelValues(...)` to avoid calling it every time.
- Add several benchmarks and it proves that storing array pointers in the map is the best. It's 6x faster than before.

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No code

Frame Graph
Before (3.6%):
![metrics_profile](https://github.com/pingcap/tiproxy/assets/29590578/31d0dc7b-9c5a-4bc8-b7b4-f4d54d4bfaff)
After (1%):
![metrics_array](https://github.com/pingcap/tiproxy/assets/29590578/db95edcc-61bf-4b68-8c45-bc3022c41fc9)

Benchmark Result:
```
BenchmarkAddCmdMetricsArrayPointer
BenchmarkAddCmdMetricsArrayPointer-8   	55810723	        22.11 ns/op
BenchmarkAddCmdMetricsArray
BenchmarkAddCmdMetricsArray-8          	 6619735	       169.0 ns/op
BenchmarkAddCmdMetricsMap
BenchmarkAddCmdMetricsMap-8            	37847948	        33.26 ns/op
BenchmarkAddCmdMetricsSimple
BenchmarkAddCmdMetricsSimple-8         	 8657454	       140.9 ns/op
```

Notable changes

- [ ] Has configuration change
- [ ] Has HTTP API interfaces change
- [ ] Has tiproxyctl change
- [ ] Other user behavior changes

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
Optimize updating metrics to improve the performance by around 3%
```